+++ packages/passport/sdk-sample-app/src/components/zkevm/ZkEvmWorkflow.tsx
@@ -1,20 +1,42 @@
-import React, { useState } from 'react';
+import React, {
+  ChangeEvent,
+  useCallback,
+  useState,
+} from 'react';
 import { Stack } from 'react-bootstrap';
 import { usePassportProvider } from '@/context/PassportProvider';
 import Request from '@/components/zkevm/Request';
 import CardStack from '@/components/CardStack';
 import { useStatusProvider } from '@/context/StatusProvider';
 import WorkflowButton from '@/components/WorkflowButton';
+import { FormControl, Toggle } from '@biom3/react';
+import { ProviderEvent } from '@imtbl/passport';
 
 function ZkEvmWorkflow() {
   const [showRequest, setShowRequest] = useState<boolean>(false);
 
-  const { isLoading } = useStatusProvider();
+  const { isLoading, addMessage } = useStatusProvider();
   const { connectZkEvm, zkEvmProvider } = usePassportProvider();
 
   const handleRequest = () => {
     setShowRequest(true);
   };
+
+  const zkEvmEventHandler = useCallback((eventName: string) => (args: any[]) => {
+    addMessage(`Provider Event: ${eventName}`, args);
+  }, [addMessage]);
+
+  const onHandleEventsChanged = useCallback((event: ChangeEvent<HTMLInputElement>) => {
+    if (event.target.checked) {
+      Object.values(ProviderEvent).forEach((eventName) => {
+        zkEvmProvider?.on(eventName, zkEvmEventHandler(eventName));
+      });
+    } else {
+      Object.values(ProviderEvent).forEach((eventName) => {
+        zkEvmProvider?.removeListener(eventName, zkEvmEventHandler(eventName));
+      });
+    }
+  }, [zkEvmEventHandler, zkEvmProvider]);
 
   return (
     <CardStack title="ZkEvm Workflow">
@@ -30,10 +52,14 @@ function ZkEvmWorkflow() {
             {showRequest
               && (
                 <Request
-                  showRequest={showRequest}
-                  setShowRequest={setShowRequest}
+                  showModal={showRequest}
+                  setShowModal={setShowRequest}
                 />
               )}
+            <FormControl sx={{ alignItems: 'center' }}>
+              <Toggle onChange={onHandleEventsChanged} />
+              <FormControl.Label>Log out events</FormControl.Label>
+            </FormControl>
           </>
         )}
         {!zkEvmProvider && (