+++ packages/passport/sdk/src/config/config.ts
@@ -50,9 +50,12 @@ export class PassportConfiguration {
 
   readonly multiRollupConfig: MultiRollupAPIConfiguration;
 
+  readonly crossSdkBridgeEnabled: boolean;
+
   constructor({
     baseConfig,
     overrides,
+    crossSdkBridgeEnabled,
     ...oidcConfiguration
   }: PassportModuleConfiguration) {
     validateConfiguration(oidcConfiguration, [
@@ -62,6 +65,8 @@ export class PassportConfiguration {
     ]);
     this.oidcConfiguration = oidcConfiguration;
     this.baseConfig = baseConfig;
+    this.crossSdkBridgeEnabled = crossSdkBridgeEnabled || false;
+
     if (overrides) {
       validateConfiguration(
         overrides,