+++ packages/passport/sdk/src/guardian/guardian.ts
@@ -1,20 +1,18 @@
 import * as guardian from '@imtbl/guardian';
-import {
-  TransactionApprovalRequestChainTypeEnum,
-  TransactionEvaluationResponse,
-} from '@imtbl/guardian';
+import { TransactionApprovalRequestChainTypeEnum, TransactionEvaluationResponse } from '@imtbl/guardian';
 import { BigNumber, ethers } from 'ethers';
 import { ConfirmationScreen } from '../confirmation';
 import { retryWithDelay } from '../network/retry';
 import { JsonRpcError, RpcErrorCode } from '../zkEvm/JsonRpcError';
 import { MetaTransaction } from '../zkEvm/types';
 import { UserZkEvm } from '../types';
+import { PassportConfiguration } from '../config';
 
 export type GuardianClientParams = {
   accessToken: string;
-  imxPublicApiDomain: string;
   confirmationScreen: ConfirmationScreen;
   imxEtherAddress: string;
+  config: PassportConfiguration;
 };
 
 export type GuardianValidateParams = {
@@ -27,6 +25,9 @@ type GuardianEVMValidationParams = {
   user: UserZkEvm;
   metaTransactions: MetaTransaction[];
 };
+
+const transactionRejectedCrossSdkBridgeError = 'Transaction requires confirmation but this functionality is not'
+  + ' supported in this environment. Please contact Immutable support if you need to enable this feature.';
 
 export const convertBigNumberishToString = (
   value: ethers.BigNumberish,
@@ -61,20 +62,23 @@ export default class GuardianClient {
   // TODO: ID-977, make this rollup agnostic
   private readonly imxEtherAddress: string;
 
+  private readonly crossSdkBridgeEnabled: boolean;
+
   constructor({
-    imxPublicApiDomain,
     accessToken,
     confirmationScreen,
     imxEtherAddress,
+    config,
   }: GuardianClientParams) {
     this.confirmationScreen = confirmationScreen;
     this.transactionAPI = new guardian.TransactionsApi(
       new guardian.Configuration({
         accessToken,
-        basePath: imxPublicApiDomain,
+        basePath: config.imxPublicApiDomain,
       }),
     );
     this.imxEtherAddress = imxEtherAddress;
+    this.crossSdkBridgeEnabled = config.crossSdkBridgeEnabled;
   }
 
   /**
@@ -96,8 +100,7 @@ export default class GuardianClient {
       this.confirmationScreen.loading(popupWindowSize);
 
       try {
-        const result = await task();
-        return result;
+        return await task();
       } catch (err) {
         this.confirmationScreen.closeWindow();
         throw err;
@@ -109,7 +112,7 @@ export default class GuardianClient {
     return this.withConfirmationScreenTask()(task);
   }
 
-  public async validate({ payloadHash }: GuardianValidateParams) {
+  public async validate({ payloadHash }: GuardianValidateParams): Promise<void> {
     const finallyFn = () => {
       this.confirmationScreen.closeWindow();
     };
@@ -135,7 +138,11 @@ export default class GuardianClient {
 
     const { confirmationRequired } = evaluateStarkexRes.data;
     if (confirmationRequired) {
-      const confirmationResult = await this.confirmationScreen.startGuardianTransaction(
+      if (this.crossSdkBridgeEnabled) {
+        throw new Error(transactionRejectedCrossSdkBridgeError);
+      }
+
+      const confirmationResult = await this.confirmationScreen.requestConfirmation(
         payloadHash,
         this.imxEtherAddress,
         TransactionApprovalRequestChainTypeEnum.Starkex,
@@ -197,8 +204,15 @@ export default class GuardianClient {
     });
 
     const { confirmationRequired, transactionId } = transactionEvaluationResponse;
+    if (confirmationRequired && this.crossSdkBridgeEnabled) {
+      throw new JsonRpcError(
+        RpcErrorCode.TRANSACTION_REJECTED,
+        transactionRejectedCrossSdkBridgeError,
+      );
+    }
+
     if (confirmationRequired && !!transactionId) {
-      const confirmationResult = await this.confirmationScreen.startGuardianTransaction(
+      const confirmationResult = await this.confirmationScreen.requestConfirmation(
         transactionId,
         this.imxEtherAddress,
         TransactionApprovalRequestChainTypeEnum.Evm,
@@ -207,8 +221,8 @@ export default class GuardianClient {
 
       if (!confirmationResult.confirmed) {
         throw new JsonRpcError(
-          RpcErrorCode.USER_REJECTED_REQUEST,
-          'Transaction rejected by user ',
+          RpcErrorCode.TRANSACTION_REJECTED,
+          'Transaction rejected by user',
         );
       }
     } else {