+++ packages/passport/sdk/src/starkEx/passportImxProvider.ts
@@ -20,7 +20,7 @@ import {
 import { ImmutableXClient } from '@imtbl/immutablex-client';
 import { IMXProvider } from '@imtbl/provider';
 import GuardianClient from '../guardian/guardian';
-import { UserImx } from '../types';
+import { PassportEventMap, PassportEvents, UserImx } from '../types';
 import { PassportError, PassportErrorType } from '../errors/passportError';
 import {
   batchNftTransfer,
@@ -31,19 +31,27 @@ import {
   createTrade,
 } from './workflows';
 import { ConfirmationScreen } from '../confirmation';
+import { PassportConfiguration } from '../config';
+import TypedEventEmitter from '../typedEventEmitter';
 
-export type PassportImxProviderInput = {
+export interface PassportImxProviderInput {
   user: UserImx;
   starkSigner: StarkSigner;
   immutableXClient: ImmutableXClient;
   confirmationScreen: ConfirmationScreen;
-  imxPublicApiDomain: string;
+  config: PassportConfiguration;
+  passportEventEmitter: TypedEventEmitter<PassportEventMap>;
+}
+
+type LoggedInPassportImxProvider = {
+  user: UserImx;
+  starkSigner: StarkSigner;
 };
 
 export class PassportImxProvider implements IMXProvider {
-  private readonly user: UserImx;
+  protected user?: UserImx;
 
-  private readonly starkSigner: StarkSigner;
+  protected starkSigner?: StarkSigner;
 
   private readonly immutableXClient: ImmutableXClient;
 
@@ -55,24 +63,43 @@ export class PassportImxProvider implements IMXProvider {
     user,
     starkSigner,
     immutableXClient,
-    imxPublicApiDomain,
     confirmationScreen,
+    config,
+    passportEventEmitter,
   }: PassportImxProviderInput) {
     this.user = user;
     this.starkSigner = starkSigner;
     this.immutableXClient = immutableXClient;
     this.confirmationScreen = confirmationScreen;
     this.guardianClient = new GuardianClient({
-      imxPublicApiDomain,
       accessToken: user.accessToken,
       confirmationScreen,
       imxEtherAddress: user.imx.ethAddress,
+      config,
     });
+
+    passportEventEmitter.on(PassportEvents.LOGGED_OUT, this.handleLogout);
+  }
+
+  private handleLogout = (): void => {
+    this.user = undefined;
+    this.starkSigner = undefined;
+  };
+
+  private checkIsLoggedIn(): asserts this is LoggedInPassportImxProvider {
+    if (this.user === undefined || this.starkSigner === undefined) {
+      throw new PassportError(
+        'User has been logged out',
+        PassportErrorType.NOT_LOGGED_IN_ERROR,
+      );
+    }
   }
 
   async transfer(
     request: UnsignedTransferRequest,
   ): Promise<CreateTransferResponseV1> {
+    this.checkIsLoggedIn();
+
     return transfer({
       request,
       user: this.user,
@@ -101,6 +128,8 @@ export class PassportImxProvider implements IMXProvider {
   }
 
   createOrder(request: UnsignedOrderRequest): Promise<CreateOrderResponse> {
+    this.checkIsLoggedIn();
+
     return createOrder({
       request,
       user: this.user,
@@ -113,6 +142,8 @@ export class PassportImxProvider implements IMXProvider {
   cancelOrder(
     request: GetSignableCancelOrderRequest,
   ): Promise<CancelOrderResponse> {
+    this.checkIsLoggedIn();
+
     return cancelOrder({
       request,
       user: this.user,
@@ -123,6 +154,8 @@ export class PassportImxProvider implements IMXProvider {
   }
 
   createTrade(request: GetSignableTradeRequest): Promise<CreateTradeResponse> {
+    this.checkIsLoggedIn();
+
     return createTrade({
       request,
       user: this.user,
@@ -135,6 +168,8 @@ export class PassportImxProvider implements IMXProvider {
   batchNftTransfer(
     request: NftTransferDetails[],
   ): Promise<CreateTransferResponse> {
+    this.checkIsLoggedIn();
+
     return batchNftTransfer({
       request,
       user: this.user,
@@ -147,6 +182,8 @@ export class PassportImxProvider implements IMXProvider {
   exchangeTransfer(
     request: UnsignedExchangeTransferRequest,
   ): Promise<CreateTransferResponseV1> {
+    this.checkIsLoggedIn();
+
     return exchangeTransfer({
       request,
       user: this.user,
@@ -188,6 +225,8 @@ export class PassportImxProvider implements IMXProvider {
   }
 
   getAddress(): Promise<string> {
+    this.checkIsLoggedIn();
+
     return Promise.resolve(this.user.imx.ethAddress);
   }
 }