+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- @imtbl/passport: Improved silent logout functionality - zkEvm provider will now emit `accountsChanged` and IMXProvider will throw an error if workflows are called after logout.
+- @imtbl/passport: Added SDK bridge support for native applications.
+
 ### Changed
 
 - @imtbl/passport: Removed `zkEvmChainId` from Passport configuration. Passport will now automatically resolve the chain ID.